+++ R/read.R
@@ -45,3 +45,49 @@ gr_read_rean <- function(file_prec, file_temp){
               pts = pts)
   )
 }
+
+#' Read ERA5 daily reanalysis data
+#' 
+#' The function reads meteorological variables (temperature and precipitation) from [ERA 5 post-processed daily statistics on single levels](https://cds.climate.copernicus.eu/datasets/derived-era5-single-levels-daily-statistics?tab=download) for using with [gr_join_rean()].
+#'
+#' @param file_prec Character string path to daily sum precipitation NetCDF file.
+#' @param file_temp Character string path to daily mean temperature NetCDF file.
+#'
+#' @return `list` containing time series, precipitation series, temperature series 
+#'   and spatial points (sf)
+#' @export
+#'
+#' @example inst/examples/gr_read_era5_daily.R
+#' 
+gr_read_era5_daily <- function(file_prec, file_temp){
+  
+  rlang::check_installed(c("ncdf4", "sf"), reason = "to use `gr_read_rean()`")
+  
+  # Read NetCDF data
+  precip = ncdf4::nc_open(file_prec)
+  temps = ncdf4::nc_open(file_temp)
+  prate = ncdf4::ncvar_get(precip, precip$var$tp) * 86400
+  temp = ncdf4::ncvar_get(temps, temps$var$t2m) - 273.15
+  
+  # Number of days and their values
+  ndays.full = temps$dim$valid_time$len
+  vals.full = temps$dim$valid_time$vals
+  
+  # Reconstruct spatial points
+  lat = precip$dim$latitude$vals
+  lon = precip$dim$longitude$vals
+  lonidx = 1:length(lon)
+  latidx = 1:length(lat)
+  coords = expand.grid(lon, lat)
+  idx = expand.grid(lonidx, latidx)
+  data = data.frame(idx, coords)
+  colnames(data) = c('nlon', 'nlat', 'lon', 'lat')
+  
+  pts = sf::st_as_sf(data, coords = c("lon", "lat"), crs = 4326)
+  
+  return(list(vals.full = vals.full,
+              prate = prate,
+              temp = temp,
+              pts = pts)
+  )
+}